# util/blockstorage: Add `TRACE_RAII`, slightly faster -reindex-chainstate with CBufferedFile