--- a/src/streams.h
+++ b/src/streams.h
@@ -640,6 +640,11 @@ public:
         src = fileIn;
     }
 
+    [[nodiscard]] bool IsNull() const
+    {
+        return src == nullptr;
+    }
+
     ~CBufferedFile()
     {
         fclose();