--- a/src/util/trace.h
+++ b/src/util/trace.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2020-2021 The Bitcoin Core developers
+// Copyright (c) 2020-2023 The Bitcoin Core developers
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
@@ -6,6 +6,8 @@
 #define BITCOIN_UTIL_TRACE_H
 
 #ifdef ENABLE_TRACING
+
+#include <util/macros.h>
 
 #include <sys/sdt.h>
 
@@ -23,6 +25,40 @@
 #define TRACE11(context, event, a, b, c, d, e, f, g, h, i, j, k) DTRACE_PROBE11(context, event, a, b, c, d, e, f, g, h, i, j, k)
 #define TRACE12(context, event, a, b, c, d, e, f, g, h, i, j, k, l) DTRACE_PROBE12(context, event, a, b, c, d, e, f, g, h, i, j, k, l)
 
+// The RAII object calls the given operation on exit.
+// This is useful to reliably trace at any return point.
+template <typename Op>
+class TraceOnExit
+{
+    Op m_op;
+
+public:
+    TraceOnExit(TraceOnExit&&) = delete;
+    TraceOnExit(TraceOnExit const&) = delete;
+    TraceOnExit& operator=(TraceOnExit&&) = delete;
+    TraceOnExit& operator=(TraceOnExit const&) = delete;
+
+    explicit TraceOnExit(Op op) : m_op(op)
+    {
+    }
+
+    ~TraceOnExit()
+    {
+        m_op();
+    }
+};
+
+// Traces entry & exit point. This can be useful to gather statistics
+// like average runtime of a code block that has multiple exit points.
+//
+// The given trace context is called with a single argument 1 for entry
+// and 0 for the exit.
+#define TRACE_RAII(context, event)                \
+    TRACE1(context, event, 1);                    \
+    auto UNIQUE_NAME(tracer) = ::TraceOnExit([] { \
+        TRACE1(context, event, 0);                \
+    })
+
 #else
 
 #define TRACE(context, event)
@@ -38,6 +74,8 @@
 #define TRACE10(context, event, a, b, c, d, e, f, g, h, i, j)
 #define TRACE11(context, event, a, b, c, d, e, f, g, h, i, j, k)
 #define TRACE12(context, event, a, b, c, d, e, f, g, h, i, j, k, l)
+
+#define TRACE_RAII(context, event)
 
 #endif