@martinus 2023-08-06
I think most of the slowdown is due to the locking required by each `fread()` call. I did a quick benchmark by switching back to `CAutoFile` and and using `fread_unlocked` instead of `std::fread` (and call `flockfile` in the ctor and `funlockfile` before `fclose`), and the performance is practically the same as with `CBufferedFile`.

I don't mind dropping the change to `CBufferedFile`, but I think the `TRACE_RAII` macro is nice (maybe with a better name, I couldn't think of anything else...). It simplifies getting accurate measurements from parts of the code.