@l0rinc 2024-11-25
@martinus, are you still working on this? If not, I'd like to take over, I've also noticed the same during profiling...